Fixing issues with consul demo and updating start
There were some typos and issues that needed resolved with the Dockerfile. I also moved the consul-template start to the packer files repo. This will copy the file to /tmp/files/ in the image which can then be called on startup with userdata. The startup script is still in progress.